fix(goreleaser): corrected the issue without having the `main.go` file in the root
