+++ infrastructure/language/python/python.go
@@ -1,4 +1,4 @@
-package main
+package python
 
 import (
 	"errors"
@@ -7,27 +7,34 @@ import (
 	"path/filepath"
 
 	"github.com/BurntSushi/toml"
+
+	"github.com/rios0rios0/autobump/config"
 )
 
+// PyProject represents the pyproject.toml file.
 type PyProject struct {
 	Project Project `toml:"project"`
 }
 
+// Project holds the project metadata from pyproject.toml.
 type Project struct {
 	Name string `toml:"name"`
 }
 
+// Python implements the Language interface for Python projects.
 type Python struct {
-	ProjectConfig ProjectConfig
+	ProjectConfig config.ProjectConfig
 }
 
+// ErrPyprojectNotFound is returned when pyproject.toml is not found.
 var ErrPyprojectNotFound = errors.New("pyproject.toml not found")
 
+// GetProjectName returns the project name from pyproject.toml.
 func (p Python) GetProjectName() (string, error) {
 	return getPyprojectName(p.ProjectConfig)
 }
 
-func getPyprojectName(projectConfig ProjectConfig) (string, error) {
+func getPyprojectName(projectConfig config.ProjectConfig) (string, error) {
 	pyprojectTomlPath := filepath.Join(projectConfig.Path, "pyproject.toml")
 
 	_, err := os.Stat(pyprojectTomlPath)
