+++ infrastructure/provider/github/github.go
@@ -1,4 +1,4 @@
-package main
+package github
 
 import (
 	"context"
@@ -6,19 +6,74 @@ import (
 	"strings"
 
 	"github.com/go-git/go-git/v5"
-	"github.com/google/go-github/v66/github"
+	"github.com/go-git/go-git/v5/plumbing/transport"
+	"github.com/go-git/go-git/v5/plumbing/transport/http"
+	gogithub "github.com/google/go-github/v66/github"
 	log "github.com/sirupsen/logrus"
+
+	"github.com/rios0rios0/autobump/config"
+	"github.com/rios0rios0/autobump/domain"
+	gitutil "github.com/rios0rios0/autobump/infrastructure/git"
 )
 
 const expectedURLParts = 2
 
-// GitHubAdapter implements PullRequestProvider for GitHub.
-type GitHubAdapter struct{}
+// Adapter implements GitServiceAdapter for GitHub.
+type Adapter struct{}
+
+// NewAdapter creates a new GitHub adapter.
+func NewAdapter() *Adapter {
+	return &Adapter{}
+}
+
+func (a *Adapter) GetServiceType() domain.ServiceType {
+	return domain.GITHUB
+}
+
+func (a *Adapter) MatchesURL(url string) bool {
+	return strings.Contains(url, "github.com")
+}
+
+func (a *Adapter) PrepareCloneURL(url string) string {
+	return url // GitHub doesn't need URL modification
+}
+
+func (a *Adapter) ConfigureTransport() {
+	// GitHub doesn't need special transport configuration
+}
+
+func (a *Adapter) GetAuthMethods(
+	_ string, // username not used for GitHub
+	globalConfig *config.GlobalConfig,
+	projectConfig *config.ProjectConfig,
+) []transport.AuthMethod {
+	var authMethods []transport.AuthMethod
+
+	// Project access token (highest priority)
+	if projectConfig.ProjectAccessToken != "" {
+		log.Infof("Using project access token to authenticate")
+		authMethods = append(authMethods, &http.BasicAuth{
+			Username: "x-access-token",
+			Password: projectConfig.ProjectAccessToken,
+		})
+	}
+
+	// GitHub personal access token
+	if globalConfig.GitHubAccessToken != "" {
+		log.Infof("Using GitHub access token to authenticate")
+		authMethods = append(authMethods, &http.BasicAuth{
+			Username: "x-access-token",
+			Password: globalConfig.GitHubAccessToken,
+		})
+	}
+
+	return authMethods
+}
 
 // PullRequestExists checks if a pull request already exists for the given source branch.
-func (g *GitHubAdapter) PullRequestExists(
-	globalConfig *GlobalConfig,
-	projectConfig *ProjectConfig,
+func (a *Adapter) PullRequestExists(
+	globalConfig *config.GlobalConfig,
+	projectConfig *config.ProjectConfig,
 	repo *git.Repository,
 	sourceBranch string,
 ) (bool, error) {
@@ -32,7 +87,7 @@ func (g *GitHubAdapter) PullRequestExists(
 	}
 
 	ctx := context.Background()
-	client := github.NewClient(nil).WithAuthToken(accessToken)
+	client := gogithub.NewClient(nil).WithAuthToken(accessToken)
 
 	// Get the repository owner and name
 	owner, repoName, err := getGitHubRepoInfo(repo)
@@ -41,7 +96,7 @@ func (g *GitHubAdapter) PullRequestExists(
 	}
 
 	// List open pull requests for the source branch
-	prs, _, err := client.PullRequests.List(ctx, owner, repoName, &github.PullRequestListOptions{
+	prs, _, err := client.PullRequests.List(ctx, owner, repoName, &gogithub.PullRequestListOptions{
 		Head:  fmt.Sprintf("%s:%s", owner, sourceBranch),
 		State: "open",
 	})
@@ -59,9 +114,9 @@ func (g *GitHubAdapter) PullRequestExists(
 }
 
 // CreatePullRequest creates a new pull request on GitHub.
-func (g *GitHubAdapter) CreatePullRequest(
-	globalConfig *GlobalConfig,
-	projectConfig *ProjectConfig,
+func (a *Adapter) CreatePullRequest(
+	globalConfig *config.GlobalConfig,
+	projectConfig *config.ProjectConfig,
 	repo *git.Repository,
 	sourceBranch string,
 	newVersion string,
@@ -76,7 +131,7 @@ func (g *GitHubAdapter) CreatePullRequest(
 	}
 
 	ctx := context.Background()
-	client := github.NewClient(nil).WithAuthToken(accessToken)
+	client := gogithub.NewClient(nil).WithAuthToken(accessToken)
 
 	// Get the repository owner and name
 	owner, repoName, err := getGitHubRepoInfo(repo)
@@ -88,7 +143,7 @@ func (g *GitHubAdapter) CreatePullRequest(
 	targetBranch := "main"
 	maintainerCanModify := true
 
-	pullRequestOptions := &github.NewPullRequest{
+	pullRequestOptions := &gogithub.NewPullRequest{
 		Title:               &prTitle,
 		Head:                &sourceBranch,
 		Base:                &targetBranch,
@@ -106,7 +161,7 @@ func (g *GitHubAdapter) CreatePullRequest(
 
 // getGitHubRepoInfo extracts owner and repository name from the remote URL.
 func getGitHubRepoInfo(repo *git.Repository) (string, string, error) {
-	remoteURL, err := getRemoteRepoURL(repo)
+	remoteURL, err := gitutil.GetRemoteRepoURL(repo)
 	if err != nil {
 		return "", "", err
 	}
